first cut at streaming results functionality
This commit adds a /subscribe endpoint which behaves similarly to the
/query endpoint, but allows the caller to pass a CallbackURL which
Pilosa will POST results to as it gets them. This is designed for the
bicliques use case where you have a long running command that will
discover results over time, and so can send those results back
asynchronously without having to maintain an open connection to the
client.