fix adding extra 0 valued bicliques in Add
the Bicliques.Add method should merge two lists of bicliques while
squashing duplicates. Due to a bug which was allocating a slice of the
appropriate length, and then using append, the result always had a bunch
of zero valued bicliques in it. This fix is to insert bicliques at the
appropriate index rather than appending them.